Introduce new top-level `job` command group
Make all job related commands, previously top-level, subcommands of the
new command group. This is intended to make the CLI more organized. The
downside of this is that it might become too "convoluted". Although it
should be more user friendly (in some sense) to introduce new non-job
related functionality in the future by introducing new command groups.